--- feature/add-task/src/main/java/team/noweekend/feature/addtask/model/AddTaskType.kt
@@ -1,9 +1,0 @@
-package team.noweekend.feature.addtask.model
-
-enum class AddTaskType(val tag: String) {
-    COMPANY(tag = "회사"),
-    PERSONAL(tag = "개인"),
-    OTHER(tag = "기타"),
-    VACATION(tag = "연차"),
-    ;
-}
